Implementing persistence of Orders in the database

+++ DeliveryApp.Data/Models/Customer.cs
@@ -4,12 +4,12 @@ namespace DeliveryApp.Data.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public long PhoneNumber { get; set; }
+        public int PhoneNumber { get; set; }
         public string Address { get; set; }
 
         public string[] ToArrString()
         {
             return new string[]{ Name , PhoneNumber.ToString(), Address };
         }
-    }    
+    }
 }
